fix: tidy up bazel vs ts types
